:recycle: extract menstruation history card state
